Redirect after login if user is admin, visual changes
Use the new GET request to redirect the admin to admin dashboard on login, or redirect the user to the home page on login. Removed unnecessary welcome text on multiple pages and replaced with something more appropriate. Also changed some layouts to look nicer and be more consistent with the rest of site #story[317]