Created seperate jest test LiveBackEndApi.test.js that tests the API calls in Api.js.  It uses a live back end with an in-memory database (not using mocking).  It isn't run when calling 'npm run unit-test'. To Run: - Start the server with in-memory database by calling './gradlew bootRun -PspringProfile=local' - In the client call 'npm run rest-test'.
NOTE: On subsequent runs of the test, the "Register a new user" test will fail because the user is already in the
database (not exactly a stateless test).  For all tests to pass you must restart the server each time.

NOTE: when running the in-memory database you may need to remove the
environment variable SPRING_DATASOURCE_URL and then use './gradlew bootRun -PspringProfile=localDev' when running the server normally from your home machine.  This is because SPRING_DATASOURCE_URL conflicts with the value in application-local.properties

Other Changes:
- Renamed "api.updateEmails" to "api.setAdditionalEmails" and
  "api.putEmails" to "api.setEmails"
- In Jest config in package.json I added a setup.js file to avoid a
  network error, and removed "jestSonar" that was giving an warning.

I'm not sure this relates to story 1, but don't know where else to put it.  #story[365]